Create pizzas gatsby's file system route api #4
