+++ src/pages/{wpPost.slug}.js
@@ -1,0 +1,38 @@
+import React from "react"
+import { graphql } from "gatsby"
+
+const BlogPostPage = ({ data }) => <h1>{data.post.title}</h1>
+
+export const pageQuery = graphql`
+  query BlogPostById(
+    # these variables are passed in via createPage.pageContext in gatsby-node.js
+    $id: String!
+  ) {
+    # selecting the current post by id
+    post: wpPost(id: { eq: $id }) {
+      slug
+      id
+      excerpt
+      content
+      title
+      date(formatString: "MMMM DD, YYYY")
+      featuredImage {
+        node {
+          altText
+          localFile {
+            childImageSharp {
+              fluid(maxWidth: 1000, quality: 100) {
+                ...GatsbyImageSharpFluid_tracedSVG
+              }
+            }
+          }
+        }
+      }
+    }
+
+
+  }
+`
+
+
+export default BlogPostPage
