+++ src/templates/blog-post.js
@@ -1,8 +1,8 @@
-import React from "react"
-import { Link, graphql } from "gatsby"
-import Img from "gatsby-image"
-import SEO from "../components/seo"
-//import parse from "html-react-parser"
+// import React from "react"
+// import { Link, graphql } from "gatsby"
+// import Img from "gatsby-image"
+// import SEO from "../components/seo"
+// import parse from "html-react-parser"
 
 // We're using Gutenberg so we need the block styles
 // import "@wordpress/block-library/build-style/style.css"
@@ -10,67 +10,67 @@ import SEO from "../components/seo"
 // import Bio from "../components/bio"
 // import Layout from "../components/layout"
 
-const BlogPostTemplate = ({ data: { previous, next, post } }) => {
-  const featuredImage = {
-    fluid: post.featuredImage?.node?.localFile?.childImageSharp?.fluid,
-    alt: post.featuredImage?.node?.alt || ``,
-  }
+// const BlogPostTemplate = ({ data: { previous, next, post } }) => {
+//   const featuredImage = {
+//     fluid: post.featuredImage?.node?.localFile?.childImageSharp?.fluid,
+//     alt: post.featuredImage?.node?.alt || ``,
+//   }
 
-  return (
-    <>
-      <SEO title="In The Morning" />
-        <div>
-          <h1>{post.title}</h1>
-          <Img  fluid={post.featuredImage.node.localFile.childImageSharp.fluid} />
-          <p dangerouslySetInnerHTML={{ __html: post.content }} />
-        </div>
-    </>
-  )
-}
+//   return (
+//     <>
+//       <SEO title="In The Morning" />
+//         <div>
+//           <h1>{post.title}</h1>
+//           <Img  fluid={post.featuredImage.node.localFile.childImageSharp.fluid} />
+//           <p dangerouslySetInnerHTML={{ __html: post.content }} />
+//         </div>
+//     </>
+//   )
+// }
 
-export default BlogPostTemplate
+// export default BlogPostTemplate
 
-export const pageQuery = graphql`
-  query BlogPostById(
-    # these variables are passed in via createPage.pageContext in gatsby-node.js
-    $id: String!
-    $previousPostId: String
-    $nextPostId: String
-  ) {
-    # selecting the current post by id
-    post: wpPost(id: { eq: $id }) {
-      tags {
-        nodes {
-            name
-        }
-      }
-      id
-      excerpt
-      content
-      title
-      date(formatString: "MMMM DD, YYYY")
-      featuredImage {
-        node {
-          altText
-          localFile {
-            childImageSharp {
-              fluid(maxWidth: 1000, quality: 100) {
-                ...GatsbyImageSharpFluid_tracedSVG
-              }
-            }
-          }
-        }
-      }
-    }
-    # this gets us the previous post by id (if it exists)
-    previous: wpPost(id: { eq: $previousPostId }) {
-      uri
-      title
-    }
-    # this gets us the next post by id (if it exists)
-    next: wpPost(id: { eq: $nextPostId }) {
-      uri
-      title
-    }
-  }
-`
+// export const pageQuery = graphql`
+//   query BlogPostById(
+//     # these variables are passed in via createPage.pageContext in gatsby-node.js
+//     $id: String!
+//     $previousPostId: String
+//     $nextPostId: String
+//   ) {
+//     # selecting the current post by id
+//     post: wpPost(id: { eq: $id }) {
+//       tags {
+//         nodes {
+//             name
+//         }
+//       }
+      // id
+      // excerpt
+      // content
+      // title
+      // date(formatString: "MMMM DD, YYYY")
+      // featuredImage {
+      //   node {
+      //     altText
+      //     localFile {
+      //       childImageSharp {
+      //         fluid(maxWidth: 1000, quality: 100) {
+      //           ...GatsbyImageSharpFluid_tracedSVG
+      //         }
+      //       }
+      //     }
+      //   }
+      // }
+//     }
+//     # this gets us the previous post by id (if it exists)
+//     previous: wpPost(id: { eq: $previousPostId }) {
+//       uri
+//       title
+//     }
+//     # this gets us the next post by id (if it exists)
+//     next: wpPost(id: { eq: $nextPostId }) {
+//       uri
+//       title
+//     }
+//   }
+// `
